requirements.txt
streamlit
pdfplumber
pandas
python-docx
openpyxl
pillow
pdfminer.six
